Created function flatArray and initialized empty array to store flattened values.
Created and defined a function called flatArray that takes "array" as a parameter. Inside the function, I initialized an empty array called arrayResult as this array will store the final flattened version of "array". This function will extract the values from the nested arrays and then push them into the arrayResult.