Implemented recursive array flattening.
Created a for loop that iterates over each of the elements (items) in the array that was passed to flatArray. Then I used a conditional where an if statement was created to check whether the current item is an array that uses array.isArray(item). If it is an array, the function will call flatArray(item) and then flatten the nested structure by using a spread operator (...). When all the elements are proccessed, then arrayResult will be returned as a flattened array.